requirements.txt
fastapi
uvicorn
torch
Pillow
efficientnet_pytorch
python-multipart
